Refactor : 플리 좋아요 mutation 로직 query 폴더로 이동
